refactor(structure): group sources into core/, standalone/ and libretro/
Move the flat list of source files under src/ into clear module directories so
the ownership of each file is obvious:

- core/      platform-independent emulator core (was the flat src/*.rs plus
             core_emulator/emulator.rs)
- standalone/ standalone-only front-end (cli.rs, gamepad_overlay.rs)
- libretro/  unchanged libretro C API

Only module paths change (crate::x -> crate::core::x); no behavioral change.
Both standalone and libretro builds pass and run.
