+++ src/core/content_loader.rs
