+++ src/core/des_constants.rs
