+++ src/core/frame_player.rs
