+++ src/core/image_decoder.rs
