--- src/core_emulator/mod.rs
@@ -1,3 +1,0 @@
-// Core emulator module.
-
-pub mod emulator;
