+++ src/standalone/mod.rs
@@ -1,0 +1,4 @@
+// Standalone front-end modules (only compiled for the binary target).
+
+pub mod cli;
+pub mod gamepad_overlay;
